# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Callcenter",
        "version": "1.0"
    },
    "paths": {
        "/api/cliente/{cedula}": {
            "get": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Mostrar cliente por documento de identidad",
                "operationId": "App\\Http\\Controllers\\Api\\Clientes\\ClienteController::cliente",
                "parameters": [
                    {
                        "name": "cedula",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar el cliente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Ruc_Cedula": {
                                            "description": "Documento de identidad del cliente",
                                            "type": "string"
                                        },
                                        "Nombre": {
                                            "description": "Nombres del cliente",
                                            "type": "string"
                                        },
                                        "Correo": {
                                            "description": "Email del cliente",
                                            "type": "string"
                                        },
                                        "telefonos": {
                                            "description": "Teléfonos asociados al cliente",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "Telefono": {
                                                        "description": "Número de teléfono",
                                                        "type": "string"
                                                    },
                                                    "direcciones": {
                                                        "description": "Direcciones asociadas al número de teléfono",
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "Telefono": {
                                                                    "description": "Número de teléfono",
                                                                    "type": "string"
                                                                },
                                                                "Direccion": {
                                                                    "description": "Dirección",
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cliente-email/{email}": {
            "get": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Mostrar cliente por email",
                "operationId": "App\\Http\\Controllers\\Api\\Clientes\\ClienteController::clientePorEmail",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar el cliente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Ruc_Cedula": {
                                            "description": "Documento de identidad del cliente",
                                            "type": "string"
                                        },
                                        "Nombre": {
                                            "description": "Nombres del cliente",
                                            "type": "string"
                                        },
                                        "Correo": {
                                            "description": "Email del cliente",
                                            "type": "string"
                                        },
                                        "telefonos": {
                                            "description": "Teléfonos asociados al cliente",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "Telefono": {
                                                        "description": "Número de teléfono",
                                                        "type": "string"
                                                    },
                                                    "direcciones": {
                                                        "description": "Direcciones asociadas al número de teléfono",
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "Telefono": {
                                                                    "description": "Número de teléfono",
                                                                    "type": "string"
                                                                },
                                                                "Direccion": {
                                                                    "description": "Dirección",
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cliente-telefono/{telefono}": {
            "get": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Mostrar clientes por número de teléfono",
                "operationId": "App\\Http\\Controllers\\Api\\Clientes\\ClienteController::clientePorTelefono",
                "parameters": [
                    {
                        "name": "telefono",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar el cliente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Ruc_Cedula": {
                                            "description": "Documento de identidad del cliente",
                                            "type": "string"
                                        },
                                        "Nombre": {
                                            "description": "Nombres del cliente",
                                            "type": "string"
                                        },
                                        "Correo": {
                                            "description": "Email del cliente",
                                            "type": "string"
                                        },
                                        "telefonos": {
                                            "description": "Teléfonos asociados al cliente",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "Telefono": {
                                                        "description": "Número de teléfono",
                                                        "type": "string"
                                                    },
                                                    "direcciones": {
                                                        "description": "Direcciones asociadas al número de teléfono",
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "Telefono": {
                                                                    "description": "Número de teléfono",
                                                                    "type": "string"
                                                                },
                                                                "Direccion": {
                                                                    "description": "Dirección",
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}